requirements.txt
waitress
flask
pillow
qrcode